graph.tcl:
Fixed two errors:

1. 'bad logscale limits (min=-0.6,max=2) for axis "x"' error.

This is the case where the user zooms into a graph, then changes the axis to
logscale (using the axis menus) and then pops the zoom stack.  The saved
min/max coordinates are for linear scale, not log.  In blt::PopZoom, save the
logscale value along with the axis min and max in the zoom stack.  (I could try
to fix the axis limits, but I'm restoring the scale when popping. This is
easier and I think it's better that the view look the same popping the stack
as going in.  You can always change it again to logscale.) 

2. 'impossible limits (min 0.746022 >= max 0.746022) for axis "y"' error.

Successively zooming in a graph will eventually go beyond the resolution of a
double (or tclprecision) and cause the graph to generate an error complaining
about impossible limits.  Now catches errors in setting the axis -min -max
options.  If error occurs (like impossible limit) the bell is sounded and the
previous zoom level is restored.